Develop Square component
